Allegro: reorganized color unit for easier dependency resolution
