+++ libsrc/allegro/include/allegro/color.h
@@ -90,28 +90,6 @@ AL_FUNC(void, create_rgb_table, (RGB_MAP *table, AL_CONST PALETTE pal, AL_METHOD
 AL_FUNC(void, create_light_table, (COLOR_MAP *table, AL_CONST PALETTE pal, int r, int g, int b, AL_METHOD(void, callback, (int pos))));
 AL_FUNC(void, create_trans_table, (COLOR_MAP *table, AL_CONST PALETTE pal, int r, int g, int b, AL_METHOD(void, callback, (int pos))));
 AL_FUNC(void, create_color_table, (COLOR_MAP *table, AL_CONST PALETTE pal, AL_METHOD(void, blend, (AL_CONST PALETTE pal, int x, int y, RGB *rgb)), AL_METHOD(void, callback, (int pos))));
-AL_FUNC(void, create_blender_table, (COLOR_MAP *table, AL_CONST PALETTE pal, AL_METHOD(void, callback, (int pos))));
-
-typedef AL_METHOD(uint32_t, BLENDER_FUNC, (uint32_t x, uint32_t y, uint32_t n));
-
-AL_FUNC(void, set_blender_mode, (BLENDER_FUNC b15, BLENDER_FUNC b16, BLENDER_FUNC b24, int r, int g, int b, int a));
-AL_FUNC(void, set_blender_mode_ex, (BLENDER_FUNC b15, BLENDER_FUNC b16, BLENDER_FUNC b24, BLENDER_FUNC b32, BLENDER_FUNC b15x, BLENDER_FUNC b16x, BLENDER_FUNC b24x, int r, int g, int b, int a));
-
-AL_FUNC(void, set_alpha_blender, (void));
-AL_FUNC(void, set_write_alpha_blender, (void));
-AL_FUNC(void, set_trans_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_add_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_burn_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_color_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_difference_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_dissolve_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_dodge_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_hue_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_invert_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_luminance_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_multiply_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_saturation_blender, (int r, int g, int b, int a));
-AL_FUNC(void, set_screen_blender, (int r, int g, int b, int a));
 
 AL_FUNC(void, hsv_to_rgb, (float h, float s, float v, int *r, int *g, int *b));
 AL_FUNC(void, rgb_to_hsv, (int r, int g, int b, float *h, float *s, float *v));
